Completely rewrite builders module
This changes almost everything; naming schemes, how the syntax tree is
built, how separate parts of the tree reference each other, bug fixes,
etc. The worst part is that these changes are non-incremental, so
there's no clean way to split it into multiple commits. In my defense,
this project is in pre-alpha and I was planning on changing these things
for a while. The good news is this is more or less what the final
product will look like, so I can start documenting the code!